Move MeshInstance and LookAt into node/nodes/Pose
