+++ src/main/java/com/marginallyclever/ro3/apps/RO3FrameDropTarget.java
@@ -2,7 +2,7 @@ package com.marginallyclever.ro3.apps;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.apps.commands.ImportScene;
-import com.marginallyclever.ro3.node.nodes.MeshInstance;
+import com.marginallyclever.ro3.node.nodes.pose.MeshInstance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
