+++ src/main/java/com/marginallyclever/ro3/node/nodes/pose/MeshInstance.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.ro3.node.nodes;
+package com.marginallyclever.ro3.node.nodes.pose;
 
 import com.marginallyclever.convenience.Ray;
 import com.marginallyclever.convenience.helpers.MatrixHelper;
@@ -6,6 +6,7 @@ import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.apps.dialogs.MeshChooserDialog;
 import com.marginallyclever.ro3.mesh.Mesh;
 import com.marginallyclever.ro3.mesh.MeshSmoother;
+import com.marginallyclever.ro3.node.nodes.Pose;
 import com.marginallyclever.ro3.raypicking.RayHit;
 import org.json.JSONObject;
