+++ src/main/java/com/marginallyclever/ro3/raypicking/RayHit.java
@@ -1,6 +1,6 @@
 package com.marginallyclever.ro3.raypicking;
 
-import com.marginallyclever.ro3.node.nodes.MeshInstance;
+import com.marginallyclever.ro3.node.nodes.pose.MeshInstance;
 
 import javax.vecmath.Vector3d;
