+++ src/main/java/com/marginallyclever/ro3/raypicking/RayPickSystem.java
@@ -3,11 +3,10 @@ package com.marginallyclever.ro3.raypicking;
 import com.marginallyclever.convenience.Ray;
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.node.Node;
-import com.marginallyclever.ro3.node.nodes.MeshInstance;
+import com.marginallyclever.ro3.node.nodes.pose.MeshInstance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.vecmath.Matrix4d;
 import java.util.*;
 
 /**
